[week2-2] 기능 구현하기 & 테스트 통과하기
* addTodo(), removeTodo() 그 외 함수들 구현하기
* context를 하나로 합치기

state관리는 provider의 역할/
view 관리는 상위 컴포넌트 역할(상위컴포넌트들 끼리는 state를 공유할 수 있도록)
