+++ src/App.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { InputProvider, TodosProvider } from './context';
+import TodosProvider from './context/TodosProvider';
 import InputBar from './components/InputBar';
 import TodoList from './components/TodoList';
 
@@ -8,9 +8,7 @@ const App = () => (
   <div className="App">
     <p>To-do</p>
     <TodosProvider>
-      <InputProvider>
-        <InputBar />
-      </InputProvider>
+      <InputBar />
       <TodoList />
     </TodosProvider>
   </div>
