+++ src/components/InputBar.jsx
@@ -1,19 +1,18 @@
 import React, { useContext } from 'react';
 
-import { InputContext, TodosContext } from '../context';
+import { TodosContext } from '../context/TodosContext';
 import Button from './views/Button';
 import Input from './views/Input';
 
 const InputBar = () => {
-  const { addTodo } = useContext(TodosContext);
-  const { handleChange, inputValue } = useContext(InputContext);
+  const { addTodo, handleChange, inputValue } = useContext(TodosContext);
 
   return (
     <div>
       <Input
         placeholder="할 일을 입력해 주세요"
         value={inputValue}
-        onChange={(event) => { handleChange(event); }}
+        onChange={handleChange}
       />
       <Button type="submit" onClick={addTodo}>
         추가
