--- src/context/InputContext.jsx
@@ -1,7 +1,0 @@
-import { createContext } from 'react';
-
-export const initialInputState = {
-  inputValue: '',
-};
-
-export const InputContext = createContext(initialInputState);
