--- src/context/index.js
@@ -1,4 +1,0 @@
-export * from './InputContext';
-export { default as InputProvider } from './providers/InputProvider';
-export * from './TodosContext';
-export { default as TodosProvider } from './providers/TodosProvider';
